Handling of subthemes
Currently, all subthemes must be created/updated to ensure a subtheme is updated in SF if it is removed in Fia. In  future change in Fia, a removed subtheme will be detectable based on status. Code should then only update subthemes that are included or removed. Change in logic is dependent on changes in Fia. To prepare for the change I moved logic into two separate methods and added comments as placeholders. This should make swapping easier and reduce risks.
Also added mapping for plan status.